refactoring to add clear flow
