+++ TraceLens/TraceUtils/split_inference/__init__.py
@@ -1,0 +1,37 @@
+###############################################################################
+# Copyright (c) 2026 Advanced Micro Devices, Inc. All rights reserved.
+#
+# See LICENSE for license information.
+###############################################################################
+
+"""Inference trace splitting: execution roots, steady-state windows, extraction."""
+
+from .execution_roots import find_iteration_roots
+from .steady_state_window import (
+    compute_reference_pd_ratio,
+    find_steady_state_window,
+    identify_steady_state_regions,
+)
+from .trace_extraction import (
+    divide_phases_and_save,
+    extract_and_save,
+    extract_iteration,
+    extract_phases_and_save,
+    get_filename,
+    parse_range,
+    preprocess_trace,
+)
+
+__all__ = [
+    "compute_reference_pd_ratio",
+    "divide_phases_and_save",
+    "extract_and_save",
+    "extract_iteration",
+    "extract_phases_and_save",
+    "find_iteration_roots",
+    "find_steady_state_window",
+    "get_filename",
+    "identify_steady_state_regions",
+    "parse_range",
+    "preprocess_trace",
+]
